[BACKEND] - Criação do endpoint para para buscar de técico por id.

+++ src/main/java/com/valdir/helpdesk/repository/TecnicoRepository.java
@@ -3,5 +3,5 @@ package com.valdir.helpdesk.repository;
 import com.valdir.helpdesk.domain.Tecnico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TecnicoRepository extends JpaRepository<Tecnico, Long> {
+public interface TecnicoRepository extends JpaRepository<Tecnico, Integer> {
 }
